Replace SciraAI service with OpenAI
